Reinforcement/rl_test/v4(chat)_chatRL/requirements.txt
numpy==1.26.4
gymnasium==0.29.1
stable-baselines3==2.3.0
torch==2.3.1
pillow
matplotlib
